# Conflicts:
#   src/projeto/otime/urls.py
#   src/projeto/otime/views.py

+++ src/projeto/otime/views.py
@@ -15,13 +15,12 @@ def index(request):
 
 def reservarHorario(request, id):
     if request.method == "POST":
-        turma = get_object_or_404(Turma, pk = id)
-        requisicao_post = request.POST
-        sala_de_aula = get_object_or_404(SalaDeAula, pk = requisicao_post["sala_de_aula"])
-        disciplina = get_object_or_404(Disciplina, pk = requisicao_post["disciplina"])
-        professor = get_object_or_404(Professor, pk = requisicao_post["professor"])
-        slot_de_horario = SlotDeHorario(posicao = requisicao_post["posicao"], sala_de_aula = sala_de_aula, disciplina = disciplina, professor = professor, turma=turma)
-        slot_de_horario.save()
+        sala_de_aula = get_object_or_404(SalaDeAula, pk = request.POST["sala_de_aula"])
+        disciplina = get_object_or_404(Disciplina, pk = request.POST["disciplina"])
+        professor = get_object_or_404(Professor, pk = request.POST["professor"])
+        for horario in request.POST.getlist("horarios"):
+            slot_de_horario = SlotDeHorario(posicao = horario, sala_de_aula = sala_de_aula, disciplina = disciplina, professor = professor)
+            slot_de_horario.save()
 
     slots_de_horario = SlotDeHorario.objects.all()
     lista_de_salas = SalaDeAula.objects.all()
@@ -39,8 +38,21 @@ def reservarHorario(request, id):
 
 
 def salas(request):
+    form = FormSala(request.POST or None)
+    if form.is_valid():
+        form.save()
+
     lista_de_salas = SalaDeAula.objects.all()
-    return render(request,'otime/salas.html',{'lista_de_salas':lista_de_salas})
+    return render(request,'otime/salas.html',{'lista_de_salas':lista_de_salas, 'form': form})
+
+def disciplinas(request):
+    form = FormDisciplina(request.POST or None)
+
+    if form.is_valid():
+        form.save()
+
+    lista_de_disciplinas = Disciplina.objects.all()
+    return render(request,'otime/disciplinas.html',{'lista_de_disciplinas':lista_de_disciplinas, 'form': form})
 
 def professores(request):
     form =  FormProfessor(request.POST or None)
@@ -49,28 +61,6 @@ def professores(request):
 
     lista_de_professores = Professor.objects.all()
     return render(request,'otime/professores.html',{'lista_de_professores':lista_de_professores, 'form':form})
-
-def disciplinas(request):
-    lista_de_disciplinas = Disciplina.objects.all()
-    return render(request,'otime/disciplinas.html',{'lista_de_disciplinas':lista_de_disciplinas})
-
-def criar_disciplina(request):
-    form = FormDisciplina(request.POST or None)
-
-    if form.is_valid():
-        form.save()
-        return redirect('disciplinas')
-
-    return render(request, 'otime/disciplina-form.html', {'form': form})
-
-def criar_sala(request):
-    form = FormSala(request.POST or None)
-
-    if form.is_valid():
-        form.save()
-        return redirect('salas')
-
-    return render(request, 'otime/sala-form.html', {'form': form})
 
 ##View para exibir as turmas cadastradas e
 ##selecionar a que irá ser definida os horarios na view de reserva de horario.
@@ -94,17 +84,11 @@ def atualizar_professor(request, id):
         form.save()
         return redirect('professores')
 
-    return render(request, 'otime/professores.html', {'form': form, 'professor': professor})
+    if request.method == 'POST':
+        professor.delete()
+        return redirect('professores')    
 
-def atualizar_disciplina(request, id):
-    disciplina = Disciplina.objects.get(id=id)
-    form =  FormDisciplina(request.POST or None, instance=disciplina)
-
-    if form.is_valid():
-        form.save()
-        return redirect('disciplinas')
-
-    return render(request, 'otime/disciplina-form.html', {'form': form, 'disciplina': disciplina})
+    return render(request, 'otime/modais/editar-prof.html', {'form': form, 'professor': professor})
 
 def atualizar_sala(request, id):
     sala = SalaDeAula.objects.get(id=id)
@@ -114,34 +98,25 @@ def atualizar_sala(request, id):
         form.save()
         return redirect('salas')
 
-    return render(request, 'otime/sala-form.html', {'form': form, 'sala': sala})
-
-def deletar_professor(request, id):
-    professor = Professor.objects.get(id=id)
-
     if request.method == 'POST':
-        professor.delete()
-        return redirect('professores')
+        sala.delete()
+        return redirect('salas')
 
-    return render(request, 'otime/prof-delete-confirm.html', {'professor': professor})
+    return render(request, 'otime/modais/editar-sala.html', {'form': form, 'sala': sala})
 
-def deletar_disciplina(request, id):
+def atualizar_disciplina(request, id):
     disciplina = Disciplina.objects.get(id=id)
+    form =  FormDisciplina(request.POST or None, instance=disciplina)
+
+    if form.is_valid():
+        form.save()
+        return redirect('disciplinas')
 
     if request.method == 'POST':
         disciplina.delete()
         return redirect('disciplinas')
 
-    return render(request, 'otime/disc-delete-confirm.html', {'disciplina': disciplina})
-
-def deletar_sala(request, id):
-    sala = SalaDeAula.objects.get(id=id)
-
-    if request.method == 'POST':
-        sala.delete()
-        return redirect('salas')
-
-    return render(request, 'otime/sala-delete-confirm.html', {'sala': sala})
+    return render(request, 'otime/modais/editar-disc.html', {'form': form, 'disciplina': disciplina})
 
 def modelo(request):
     lista_de_professores = Professor.objects.all()
